comment out some code as I will be moving it onto a separate branch
* You probably have 2 usages for redis as a queue:
  - Standalone instances
  - Master/slave setup (master for writes/reads, slaves for backups)

So in the second approach we probably don't want to monitor the slaves
unless they become masters.

I want to control failovers and stop monitoring on nodes that become
slaves but I'll do that on a separate dev branch.

For the moment the master branch can be used to easily monitor
standalone instances.